requirements.txt
anyio==4.0.0
asttokens==2.4.1
certifi==2023.7.22
charset-normalizer==3.3.2
decorator==5.1.1
exceptiongroup==1.1.3
executing==2.0.1
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
ipython==8.17.2
jedi==0.19.1
matplotlib-inline==0.1.6
numpy==1.26.1
pandas==2.1.2
parso==0.8.3
pexpect==4.8.0
prompt-toolkit==3.0.39
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.16.1
python-dateutil==2.8.2
python-telegram-bot==20.6
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
stack-data==0.6.3
traitlets==5.13.0
tzdata==2023.3
urllib3==2.0.7
wcwidth==0.2.9
